@@ -1,6 +1,80 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "21cmMC"
+description = "An extensible MCMC framework for 21cmFAST."
+authors = [
+    {name="Brad Greig", email="greigb@unimelb.edu.au"},
+    {name="Steven Murray", email='steven.murray@sns.it'},
+]
+readme = "README.rst"
+requires-python = ">=3.8"
+license = {text = "MIT"}
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: Unix",
+    "Operating System :: POSIX",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: Implementation :: CPython",
+]
+dependencies = [
+    "click",
+    "numpy",
+    "cosmoHammer",
+    "scipy",
+    "matplotlib>=2.1",
+    "emcee<3",
+    "powerbox>=0.5.7",
+    "cached_property",
+    "21cmFAST",
+    "pymultinest",
+    "ultranest",
+    "py21cmemu>=1.0.8",
+    "astropy>=5.2.1",
+]
+dynamic = [
+    "version",
+]
+
+[project.optional-dependencies]
+samplers = [
+  "pymultinest",
+  "ultranest",
+]
+test = [
+  "pytest",
+  "pytest-cov",
+]
+docs = [
+  "ipython",
+  "nbsphinx",
+  "numpydoc",
+  "sphinx>=1.3",
+  "furo",
+]
+dev = [
+  "pre-commit",
+  "21cmMC[test,docs]"
+]
+
+[tool.setuptools_scm]
+write_to = "src/py21cmmc/_version.py"
+parentdir_prefix_version = "21CMMC-"
+fallback_version = "0.0.0"
+
+# [tool.setuptools]
+# packages = ["21CMMC"]
+
 [tool.black]
 include = '\.pyi?$'
-target-version = ['py36', 'py37']
+target-version = ['py38', 'py39', 'py310', 'py311']
 exclude = '''
 /(
     \.git
@@ -23,3 +97,21 @@ known_first_party = 'py21cmmc'
 sections = ['FUTURE','STDLIB','COMPAT','TEST','THIRDPARTY','FIRSTPARTY','LOCALFOLDER']
 default_section = 'THIRDPARTY'
 profile = 'black'
+
+[tool.pytest.ini_options]
+testpaths = "tests"
+norecursedirs = [
+  "migrations",
+]
+python_files = [
+  "test_*.py",
+  "*_test.py",
+  "tests.py",
+]
+addopts = [
+  "-ra",
+  "--strict",
+  "--doctest-modules",
+  "--doctest-glob=\\*.rst",
+  "--tb=short",
+]
